Fix docs, remove unused structs
