+++ amethyst_assets/src/asset.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use {BoxedErr, Source, SourceId};
+use {BoxedErr, Source};
 
 /// One of the three core traits of this crate.
 ///
@@ -19,28 +19,6 @@ use {BoxedErr, Source, SourceId};
 pub trait Asset: Send + Sync + 'static {
     /// The `Data` type the asset can be created from.
     type Data: Send + Sync + 'static;
-}
-
-/// A specifier for an asset, uniquely identifying it by
-///
-/// * the extension (the format it was provided in)
-/// * its name
-/// * the storage it was loaded from
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct AssetSpec {
-    /// The possible extensions of this asset
-    pub exts: &'static [&'static str],
-    /// The name of this asset.
-    pub name: String,
-    /// Unique identifier indicating the Storage from which the asset was loaded.
-    pub store: SourceId,
-}
-
-impl AssetSpec {
-    /// Creates a new asset specifier from the given parameters.
-    pub fn new(name: String, exts: &'static [&'static str], store: SourceId) -> Self {
-        AssetSpec { exts, name, store }
-    }
 }
 
 /// A format, providing a conversion from bytes to asset data, which is then
