+++ amethyst_assets/src/reload.rs
@@ -1,3 +1,5 @@
+//! MODULE DISABLED, will be reused later!
+
 use Source;
 
 pub trait Reload {
